Use concerns and draper to dry up code
