+++ app/controllers/concerns/reroute_manager.rb
@@ -1,4 +1,6 @@
-module RerouteHelper
+module RerouteManager
+  include ApplicationHelper
+
   def manage_redirection(target)
     #Do all the statistics
     if target.active
@@ -8,13 +10,17 @@ module RerouteHelper
     end
   end
 
-  def manage_inactive_target(url)
+  def set_url_deactivated_message(url)
     flash[:error] = "The url: #{host_url+url} has been deactivated by the owner."
-    redirect_to root_path
+    manage_inactive_target
   end
 
   def manage_active_target(target)
     target.save_this_visit
     redirect_to (target.original)
+  end
+
+  def manage_inactive_target
+    current_user ? redirect_to dashboard_url : redirect_to root_path
   end
 end
