+++ app/controllers/sessions_controller.rb
@@ -14,6 +14,7 @@ class SessionsController < ApplicationController
 
   def destroy
     session[:user_id] = nil
-    redirect_to root_url, notice: "Signed out"
+    flash[:notice] = "Signed out"
+    redirect_to root_url
   end
 end
