+++ app/controllers/users_controller.rb
@@ -2,6 +2,7 @@ class UsersController < ApplicationController
   before_action :authenticate_user
 
   def index
-    @urls = current_user.get_urls
+    @user_decorator = current_user
+    # @urls = current_user.get_urls.decorate
   end
 end
