+++ app/decorators/user_decorator.rb
@@ -1,0 +1,9 @@
+class UserDecorator < Draper::Decorator
+  delegate_all
+
+  def get_user_token
+    token
+  end
+
+
+end
