+++ app/helpers/application_helper.rb
@@ -1,2 +1,5 @@
 module ApplicationHelper
+  def host_url
+    request.base_url + "/"
+  end
 end
