+++ spec/decorators/user_decorator_spec.rb
@@ -1,0 +1,4 @@
+require 'spec_helper'
+
+describe UserDecorator do
+end
